Move NavigateToItemProviderFactory to VS layer
